backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
pyjwt==2.8.0
pyyaml==6.0.1
pypdf2==3.0.1
qrcode[pil]==7.4.2
python-jose[cryptography]==3.3.0
slowapi==0.1.9
mangum==0.17.0
